2020.3.28 v1.2 的多种修改，删除旧代码，单元重新归类，单元测试调整，基本完成；
